Create To-Do List with GUI
Created a To-Do List in which the user has a GUI to work with. The user can add new items to the list and can also delete them from the list. Every time a new change is made the list gets a refresh in order for the actions to be more fluent. There is also a "View Tasks" window in order to see what tasks remain and to delete the ones that have been completed.