Create Calculator with GUI
I made a calculator that has GUI and keyboard functionality.

But when pressing "Enter" or "C" the keyboard input wont work as I want it to work. When pressing "Enter" I want it to show the results, when pressing "C" I want it to delete the entire calculation. 

Will figure it out later.